requirements.txt
aenum==3.1.11
allure-python-commons==2.9.45
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
delayed-assert==0.3.6
docopt==0.6.2
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
jsonschema==4.6.0
packaging==21.3
page==0.1.7
pipreqs==0.4.11
playwright==1.22.0
pluggy==1.0.0
py==1.11.0
pyee==8.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-assertions==0.5.0
pytest-base-url==2.0.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-playwright==0.3.0
python-slugify==6.1.2
requests==2.28.0
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.9
websockets==10.1
yarg==0.1.9
